fix(localizations): Rename to `en-UY.ts`
